+++ component/src/main/java/dev/ikm/tinkar/component/location/SpatialPoint.java
@@ -15,4 +15,4 @@
  */
 package dev.ikm.tinkar.component.location;
 
-public record SpatialPoint(int x, int y, int z) { }
+public record SpatialPoint(float x, float y, float z) { }
